refactor existing logic
